the-ams-bank-system/the-ams-bank-system/src/main/java/com/ams/theamsbanksystem/service/AccountService.java
package com.ams.theamsbanksystem.service;

import com.ams.theamsbanksystem.model.Account;
import com.ams.theamsbanksystem.repository.AccountRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;
import java.util.Optional;

@Service
public class AccountService {

    private final AccountRepository accountRepository;

    @Autowired
    public AccountService(AccountRepository accountRepository) {
        this.accountRepository = accountRepository;
    }

    public List<Account> getAllAccounts() {
        return accountRepository.findAll();
    }

    public Optional<Account> getAccountById(Long id) {
        return accountRepository.findById(id);
    }

    public Account createAccount(Account account) {
        // Add any additional logic or checks before saving the account
        return accountRepository.save(account);
    }

    public Account updateAccount(Long id, Account accountDetails) {
        Account account = accountRepository.findById(id)
                .orElseThrow(() -> new IllegalStateException("Account not found for this id: " + id));

        account.setAccountNumber(accountDetails.getAccountNumber());
        account.setBalance(accountDetails.getBalance());
        // Update other fields as necessary

        return accountRepository.save(account);
    }

    public void deleteAccount(Long id) {
        Account account = accountRepository.findById(id)
                .orElseThrow(() -> new IllegalStateException("Account not found for this id: " + id));
        accountRepository.delete(account);
    }

    // In AccountService.java

}
